Refactor tutorials and move to separate folder
