+++ Met4FoF_redundancy/__init__.py
@@ -1,7 +1,1 @@
 from .MFred import redundancy1
-from .agentMFred import (
-    metrological_streams_v2,
-    redundancyAgents_tutorial_1,
-    redundancyAgents1,
-    redundancyAgents_tutorial_2,
-)
