+++ Met4FoF_redundancy_tutorials/redundancyAgents_tutorial_2.py
@@ -9,8 +9,8 @@ import numpy as np
 from agentMET4FOF.agents import AgentNetwork
 from agentMET4FOF.metrological_agents import MetrologicalMonitorAgent
 
-from .metrological_streams_v2 import MetrologicalMultiWaveGenerator
-from .redundancyAgents1 import MetrologicalMultiWaveGeneratorAgent, RedundancyAgent
+from Met4FoF_redundancy.agentMFred.metrological_streams_v2 import MetrologicalMultiWaveGenerator
+from Met4FoF_redundancy.agentMFred.redundancyAgents1 import MetrologicalMultiWaveGeneratorAgent, RedundancyAgent
 
 
 def main():
